@@ -1,29 +1,77 @@
 {
   "name": "w1l",
-  "version": "0.1.0",
+  "version": "2.2.1",
   "private": true,
+  "type": "module",
   "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build"
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview --port 5050",
+    "lint": "eslint . -c .eslintrc.cjs --fix --ext .ts,.js,.cjs,.vue,.tsx,.jsx",
+    "build:icons": "tsx src/plugins/iconify/build-icons.js",
+    "postinstall": "npm run build:icons"
   },
   "dependencies": {
-    "axios": "^1.6.0",
-    "bootstrap-icons": "^1.11.1",
-    "core-js": "^3.8.3",
-    "sweetalert2": "^11.10.1",
-    "vue": "^3.2.13",
-    "vue-router": "^4.0.3",
-    "vue-tables-2": "^2.3.5",
-    "vuetify": "^3.5.1",
-    "vuex": "^4.0.0"
+    "@floating-ui/dom": "1.5.3",
+    "@iconify-json/bxl": "^1.1.10",
+    "@vueuse/core": "^10.5.0",
+    "@vueuse/math": "^10.5.0",
+    "apexcharts": "3.37.0",
+    "axios": "^1.6.8",
+    "eslint-import-resolver-custom-alias": "^1.3.2",
+    "pinia": "^2.1.7",
+    "prismjs": "^1.29.0",
+    "roboto-fontface": "^0.10.0",
+    "sweetalert2": "^11.10.7",
+    "vue": "^3.3.4",
+    "vue-flatpickr-component": "11.0.3",
+    "vue-prism-component": "^2.0.0",
+    "vue-router": "^4.2.5",
+    "vue3-apexcharts": "^1.4.4",
+    "vue3-perfect-scrollbar": "^1.6.1",
+    "vuetify": "3.4.9",
+    "webfontloader": "^1.6.28"
   },
   "devDependencies": {
-    "@mdi/font": "^7.3.67",
-    "@vue/cli-plugin-babel": "~5.0.0",
-    "@vue/cli-plugin-router": "~5.0.0",
-    "@vue/cli-plugin-vuex": "~5.0.0",
-    "@vue/cli-service": "~5.0.0",
-    "sass": "^1.32.7",
-    "sass-loader": "^12.0.0"
-  }
+    "@antfu/eslint-config-vue": "^0.43.1",
+    "@antfu/utils": "^0.7.6",
+    "@iconify-json/ri": "^1.1.18",
+    "@iconify/tools": "^3.0.5",
+    "@iconify/utils": "^2.1.11",
+    "@iconify/vue": "4.1.1",
+    "@vitejs/plugin-vue": "^5.0.2",
+    "@vitejs/plugin-vue-jsx": "^3.0.2",
+    "eslint": "^8.56.0",
+    "eslint-config-airbnb-base": "^15.0.0",
+    "eslint-import-resolver-typescript": "^3.6.1",
+    "eslint-plugin-case-police": "^0.6.1",
+    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-promise": "^6.1.1",
+    "eslint-plugin-regex": "^1.10.0",
+    "eslint-plugin-sonarjs": "^0.23.0",
+    "eslint-plugin-unicorn": "^50.0.1",
+    "eslint-plugin-vue": "^9.19.2",
+    "postcss-html": "^1.5.0",
+    "postcss-scss": "^4.0.9",
+    "sass": "^1.69.4",
+    "stylelint": "15.11.0",
+    "stylelint-codeguide": "0.3.2",
+    "stylelint-config-idiomatic-order": "9.0.0",
+    "stylelint-config-standard-scss": "11.1.0",
+    "stylelint-use-logical-spec": "5.0.0",
+    "tsx": "^4.7.0",
+    "unplugin-auto-import": "^0.17.2",
+    "unplugin-vue-components": "^0.26.0",
+    "vite": "^5.0.5",
+    "vite-plugin-vue-layouts": "^0.11.0",
+    "vite-plugin-vuetify": "2.0.1",
+    "vite-svg-loader": "^5.1.0"
+  },
+  "resolutions": {
+    "stylelint-order": "6.0.3"
+  },
+  "overrides": {
+    "stylelint-order": "6.0.3"
+  },
+  "packageManager": "pnpm@8.6.2"
 }
