@@ -1,35 +1,22 @@
 import { createApp } from 'vue'
-import App from './App.vue'
-import router from './router'
-import store from './store'
-import './css/styles.scss'
-
-import 'vuetify/styles'
-import '@mdi/font/css/materialdesignicons.css'
-import { createVuetify } from 'vuetify'
-import * as components from 'vuetify/components'
-import * as directives from 'vuetify/directives'
-
+import App from '@/App.vue'
+import { registerPlugins } from '@core/utils/plugins'
+// import axiosConfig from './axiosConfig.js';
 import axios from 'axios';
 
 
-const customAxios = axios.create({
-    baseURL: process.env.VUE_APP_API_URL,
-    timeout: 10000,
-});
+// Styles
+import '@core/scss/template/index.scss'
+import '@layouts/styles/index.scss'
 
-const app = createApp(App);
-app.config.globalProperties.$http = customAxios;
+// Create vue app
+const app = createApp(App)
 
-const vuetify = createVuetify({
-    components,
-    directives,
-    theme: {
-      defaultTheme: 'dark'
-    },
-    icons: {
-      defaultSet: 'mdi',
-    },
-  })
+axios.defaults.baseURL = 'https://localhost:7192/api';
+app.config.globalProperties.$axios = axios;
 
-app.use(store).use(router).use(vuetify).mount('#app')
+// Register plugins
+registerPlugins(app)
+
+// Mount vue app
+app.mount('#app')
